Refactor compiled-engine ownership boundaries
