+++ beamz/simulation/compiler/__init__.py
@@ -1,0 +1,3 @@
+"""Internal compiler package for configuration-specialized simulations."""
+
+__all__: list[str] = []
