+++ beamz/simulation/compiler/runtime.py
@@ -1,0 +1,100 @@
+"""Runtime state containers for the compiled simulation engine."""
+
+from __future__ import annotations
+
+from dataclasses import dataclass
+from typing import NamedTuple
+
+import jax.numpy as jnp
+
+
+class EngineState(NamedTuple):
+    """Runtime EM field state."""
+
+    ex: jnp.ndarray
+    ey: jnp.ndarray
+    ez: jnp.ndarray
+    hx: jnp.ndarray
+    hy: jnp.ndarray
+    hz: jnp.ndarray
+    t: jnp.ndarray
+    current_step: jnp.ndarray
+
+
+class MonitorState(NamedTuple):
+    """Packed monitor accumulators."""
+
+    powers: jnp.ndarray
+    timestamps: jnp.ndarray
+    counts: jnp.ndarray
+    freq_flux_re: jnp.ndarray
+    freq_flux_im: jnp.ndarray
+    freq_phase_re: jnp.ndarray
+    freq_phase_im: jnp.ndarray
+    dft_vec_re: jnp.ndarray
+    dft_vec_im: jnp.ndarray
+    dft_weight_sum: jnp.ndarray
+
+
+class PowerMonitorState(NamedTuple):
+    """Narrow monitor state for power-only accumulation."""
+
+    powers: jnp.ndarray
+    timestamps: jnp.ndarray
+    counts: jnp.ndarray
+
+
+class SpectralMonitorState(NamedTuple):
+    """Monitor state for power + frequency accumulation without DFT tensors."""
+
+    powers: jnp.ndarray
+    timestamps: jnp.ndarray
+    counts: jnp.ndarray
+    freq_flux_re: jnp.ndarray
+    freq_flux_im: jnp.ndarray
+    freq_phase_re: jnp.ndarray
+    freq_phase_im: jnp.ndarray
+
+
+class UpdateCoefficients(NamedTuple):
+    """Static update coefficients passed as runtime arguments."""
+
+    h_decay_x: jnp.ndarray
+    h_source_x: jnp.ndarray
+    h_source_lossless_x: jnp.ndarray
+    h_decay_y: jnp.ndarray
+    h_source_y: jnp.ndarray
+    h_source_lossless_y: jnp.ndarray
+    h_decay_z: jnp.ndarray
+    h_source_z: jnp.ndarray
+    h_source_lossless_z: jnp.ndarray
+    e_decay_x: jnp.ndarray
+    e_source_x: jnp.ndarray
+    e_source_lossless_x: jnp.ndarray
+    e_decay_y: jnp.ndarray
+    e_source_y: jnp.ndarray
+    e_source_lossless_y: jnp.ndarray
+    e_decay_z: jnp.ndarray
+    e_source_z: jnp.ndarray
+    e_source_lossless_z: jnp.ndarray
+
+
+class RunState(NamedTuple):
+    """Auxiliary run counters."""
+
+    compile_count: jnp.ndarray
+
+
+@dataclass(frozen=True)
+class CompiledRunConfig:
+    """Static compiled run configuration."""
+
+    resolution: float
+    dt: float
+    num_steps: int
+    plane_2d: str
+    is_3d: bool
+    precision: str = "float32"
+    loop_kind: str = "scan"
+    source_single_slab_dense: bool = False
+    temporal_block_steps: int = 1
